requirements.txt
Django==2.2.11
python-dotenv==0.12.0
psycopg2-binary==2.8.4
django-filter==2.2.0
djangorestframework==3.11.0
django-model-utils==4.0.0
factory-boy==2.12.0
